Add missing header. Minor cleanup

+++ src/demos/vehicle/demo_MROLE/demo_VEH_MROLE.cpp
@@ -9,7 +9,7 @@
 // http://projectchrono.org/license-chrono.txt.
 //
 // =============================================================================
-// Authors: Radu Serban, Justin Madsen
+// Authors: Radu Serban, Rainer Gericke
 // =============================================================================
 //
 // Main driver function for the mrole full model.
@@ -320,7 +320,7 @@ int main(int argc, char* argv[]) {
         my_mrole.Advance(step_size);
         app.Advance(step_size);
         GetLog() << "======================================================================\n";
-        for (size_t i = 0; i < 4; i++) {
+        for (int i = 0; i < 4; i++) {
             double sp1l = my_mrole.GetVehicle().GetDriveline()->GetSpindleTorque(i, LEFT);
             double sp1r = my_mrole.GetVehicle().GetDriveline()->GetSpindleTorque(i, RIGHT);
             GetLog() << "Axle#" << i << "  Tleft = " << sp1l << "   Tright = " << sp1r << "\n";
